# shared/openapi/tedmarks.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tedmarks API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sync/pull": {
      "get": {
        "operationId": "syncPull",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Changes since serverSeq",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sync/push": {
      "post": {
        "operationId": "syncPush",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted, newer, and rejected records",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPushResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SyncPushRequest": {
        "type": "object",
        "properties": {
          "changes": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "modifiedAt": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "modifiedAt"
                ],
                "additionalProperties": {}
              }
            }
          }
        },
        "required": [
          "changes"
        ]
      },
      "SyncPushResponse": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "collection": {
                  "type": "string"
                },
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "serverSeq": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "collection",
                "id",
                "serverSeq"
              ],
              "additionalProperties": false
            }
          },
          "newer": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": {}
              }
            }
          },
          "rejected": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "collection": {
                  "type": "string"
                },
                "id": {
                  "type": "string"
                },
                "reason": {
                  "type": "string"
                }
              },
              "required": [
                "collection",
                "id",
                "reason"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "accepted",
          "newer",
          "rejected"
        ],
        "additionalProperties": false
      },
      "SyncPullResponse": {
        "type": "object",
        "properties": {
          "changes": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": {}
              }
            }
          },
          "serverSeq": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "changes",
          "serverSeq",
          "hasMore"
        ],
        "additionalProperties": false
      }
    }
  }
}